// frontend/projects/ng-advanced-grid/src/lib/interfaces/grid.interfaces.ts
export interface GridPosition {
  x: number;
  y: number;
}

export interface GridSize {
  width: number;
  height: number;
}

export interface GridBounds {
  top: number;
  left: number;
  right: number;
  bottom: number;
}

export interface GridItem {
  id: string;
  x: number;
  y: number;
  width: number;
  height: number;
  minWidth?: number;
  minHeight?: number;
  maxWidth?: number;
  maxHeight?: number;
  fixed?: boolean;
  resizable?: boolean;
  draggable?: boolean;
  data?: any;
}
// ...
export interface DragEvent {
  item: GridItem;
  position: GridPosition;
  size: GridSize;
  bounds: GridBounds;
  event: MouseEvent | TouchEvent;
}

export interface ResizeEvent {
  item: GridItem;
  position: GridPosition;
  size: GridSize;
  bounds: GridBounds;
  direction: ResizeDirection;
  event: MouseEvent | TouchEvent;
}

export type ResizeDirection = 
  | 'n' | 's' | 'e' | 'w' 
  | 'ne' | 'nw' | 'se' | 'sw';